+++ src/components/Core/RightClickMenu/index.ts
@@ -1,0 +1,3 @@
+import RightClickMenu from './RightClickMenu.vue'
+
+export default RightClickMenu
