+++ src/components/TorrentDetail/Content/index.ts
@@ -1,0 +1,3 @@
+import Content from './Content.vue'
+
+export default Content
