+++ src/pages/Dashboard.vue
@@ -1,15 +1,26 @@
 <script lang="ts" setup>
-import RightClickMenu from '@/components/Dashboard/TRC/RightClickMenu.vue'
+import RightClickMenu from '@/components/Core/RightClickMenu'
+import Toolbar from '@/components/Dashboard/Toolbar.vue'
 import GridView from '@/components/Dashboard/Views/Grid/GridView.vue'
 import ListView from '@/components/Dashboard/Views/List/ListView.vue'
 import TableView from '@/components/Dashboard/Views/Table/TableView.vue'
 import ConfirmDeleteDialog from '@/components/Dialogs/ConfirmDeleteDialog.vue'
-import Toolbar from '@/components/Dashboard/Toolbar.vue'
+import MoveTorrentDialog from '@/components/Dialogs/MoveTorrentDialog.vue'
+import RenameTorrentDialog from '@/components/Dialogs/RenameTorrentDialog.vue'
+import ShareLimitDialog from '@/components/Dialogs/ShareLimitDialog.vue'
+import SpeedLimitDialog from '@/components/Dialogs/SpeedLimitDialog.vue'
 import { useArrayPagination } from '@/composables'
 import { DashboardDisplayMode } from '@/constants/vuetorrent'
 import { doesCommand } from '@/helpers'
-import { useDashboardStore, useDialogStore, useMaindataStore, useTorrentStore, useVueTorrentStore } from '@/stores'
-import { Torrent as TorrentType } from '@/types/vuetorrent'
+import {
+  useDashboardStore,
+  useDialogStore,
+  useMaindataStore,
+  usePreferenceStore,
+  useTorrentStore,
+  useVueTorrentStore
+} from '@/stores'
+import { Torrent as TorrentType, TRCMenuEntry } from '@/types/vuetorrent'
 import { storeToRefs } from 'pinia'
 import { computed, nextTick, onBeforeMount, onBeforeUnmount, onMounted, reactive, ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
@@ -21,6 +32,7 @@ const dashboardStore = useDashboardStore()
 const { currentPage: dashboardPage, isSelectionMultiple, selectedTorrents, displayMode } = storeToRefs(dashboardStore)
 const dialogStore = useDialogStore()
 const maindataStore = useMaindataStore()
+const preferenceStore = usePreferenceStore()
 const torrentStore = useTorrentStore()
 const { filteredTorrents } = storeToRefs(torrentStore)
 const vuetorrentStore = useVueTorrentStore()
@@ -171,6 +183,262 @@ function handleKeyboardShortcuts(e: KeyboardEvent) {
   }
 }
 
+const isMultiple = computed(() => dashboardStore.selectedTorrents.length > 1)
+const hashes = computed(() => dashboardStore.selectedTorrents)
+const hash = computed(() => hashes.value[0])
+const torrent = computed(() => torrentStore.getTorrentByHash(hash.value))
+const torrents = computed(() => dashboardStore.selectedTorrents.map(torrentStore.getTorrentByHash).filter(torrent => !!torrent))
+const availableCategories = computed(() => [{ name: '' }, ...maindataStore.categories])
+
+async function resumeTorrents() {
+  await torrentStore.resumeTorrents(hashes)
+}
+
+async function forceResumeTorrents() {
+  await torrentStore.forceResumeTorrents(hashes)
+}
+
+async function pauseTorrents() {
+  await torrentStore.pauseTorrents(hashes)
+}
+
+function deleteTorrents() {
+  dialogStore.createDialog(ConfirmDeleteDialog, { hashes: [...dashboardStore.selectedTorrents] })
+}
+
+function setDownloadPath() {
+  dialogStore.createDialog(MoveTorrentDialog, { hashes: [...dashboardStore.selectedTorrents], mode: 'dl' })
+}
+
+function setSavePath() {
+  dialogStore.createDialog(MoveTorrentDialog, { hashes: [...dashboardStore.selectedTorrents], mode: 'save' })
+}
+
+function renameTorrents() {
+  dialogStore.createDialog(RenameTorrentDialog, { hash: dashboardStore.selectedTorrents[0] })
+}
+
+async function forceRecheck() {
+  await torrentStore.recheckTorrents(hashes)
+}
+
+async function forceReannounce() {
+  await maindataStore.reannounceTorrents(hashes)
+}
+
+async function toggleSeqDl() {
+  await maindataStore.toggleSeqDl(hashes)
+}
+
+async function toggleFLPiecePrio() {
+  await maindataStore.toggleFLPiecePrio(hashes)
+}
+
+async function toggleAutoTMM() {
+  await maindataStore.toggleAutoTmm(hashes, !torrent.value?.auto_tmm)
+}
+
+function hasTag(tag: string) {
+  return torrents.value.every(torrent => torrent && torrent.tags && torrent.tags.includes(tag))
+}
+
+async function toggleTag(tag: string) {
+  if (hasTag(tag)) await torrentStore.removeTorrentTags(hashes.value, [tag])
+  else await torrentStore.addTorrentTags(hashes.value, [tag])
+}
+
+async function copyValue(valueToCopy: string) {
+  try {
+    await navigator.clipboard.writeText(valueToCopy)
+  } catch (error) {
+    toast.error(t('toast.copy.error'))
+    return
+  }
+
+  toast.success(t('toast.copy.success'))
+}
+
+function setDownloadLimit() {
+  dialogStore.createDialog(SpeedLimitDialog, { hashes: hashes.value, mode: 'download' })
+}
+
+function setUploadLimit() {
+  dialogStore.createDialog(SpeedLimitDialog, { hashes: hashes.value, mode: 'upload' })
+}
+
+function setShareLimit() {
+  dialogStore.createDialog(ShareLimitDialog, { hashes: hashes.value })
+}
+
+async function exportTorrents() {
+  hashes.value.forEach(hash => {
+    torrentStore.exportTorrent(hash).then(blob => {
+      const url = window.URL.createObjectURL(blob)
+      const link = document.createElement('a')
+      link.href = url
+      link.style.opacity = '0'
+      link.setAttribute('download', `${hash}.torrent`)
+      document.body.appendChild(link)
+      link.click()
+      document.body.removeChild(link)
+    })
+  })
+}
+
+const menuData = computed<TRCMenuEntry[]>(() => [
+  {
+    text: t('dashboard.right_click.advanced.title'),
+    icon: 'mdi-head-cog',
+    children: [
+      {
+        text: t('dashboard.right_click.advanced.download_path'),
+        icon: 'mdi-tray-arrow-down',
+        action: setDownloadPath
+      },
+      {
+        text: t('dashboard.right_click.advanced.save_path'),
+        icon: 'mdi-content-save',
+        action: setSavePath
+      },
+      {
+        text: t('dashboard.right_click.advanced.rename'),
+        icon: 'mdi-rename-box',
+        hidden: isMultiple.value,
+        action: renameTorrents
+      },
+      {
+        text: t('dashboard.right_click.advanced.recheck'),
+        icon: 'mdi-playlist-check',
+        action: forceRecheck
+      },
+      {
+        text: t('dashboard.right_click.advanced.reannounce'),
+        icon: 'mdi-bullhorn',
+        action: forceReannounce
+      },
+      {
+        text: t('dashboard.right_click.advanced.seq_dl'),
+        icon: torrent.value?.seq_dl ? 'mdi-checkbox-marked' : 'mdi-checkbox-blank-outline',
+        action: toggleSeqDl
+      },
+      {
+        text: t('dashboard.right_click.advanced.f_l_prio'),
+        icon: torrent.value?.f_l_piece_prio ? 'mdi-checkbox-marked' : 'mdi-checkbox-blank-outline',
+        action: toggleFLPiecePrio
+      },
+      {
+        text: t('dashboard.right_click.advanced.auto_tmm'),
+        icon: torrent.value?.auto_tmm ? 'mdi-checkbox-marked' : 'mdi-checkbox-blank-outline',
+        action: toggleAutoTMM
+      }
+    ]
+  },
+  {
+    text: t('dashboard.right_click.priority.title'),
+    icon: 'mdi-priority-high',
+    hidden: !preferenceStore.preferences!.queueing_enabled,
+    children: [
+      {
+        text: t('dashboard.right_click.priority.top'),
+        icon: 'mdi-priority-high',
+        action: async () => await torrentStore.setTorrentPriority(hashes.value, 'topPrio')
+      },
+      {
+        text: t('dashboard.right_click.priority.increase'),
+        icon: 'mdi-arrow-up',
+        action: async () => await torrentStore.setTorrentPriority(hashes.value, 'increasePrio')
+      },
+      {
+        text: t('dashboard.right_click.priority.decrease'),
+        icon: 'mdi-arrow-down',
+        action: async () => await torrentStore.setTorrentPriority(hashes.value, 'decreasePrio')
+      },
+      {
+        text: t('dashboard.right_click.priority.bottom'),
+        icon: 'mdi-priority-low',
+        action: async () => await torrentStore.setTorrentPriority(hashes.value, 'bottomPrio')
+      }
+    ]
+  },
+  {
+    text: t('dashboard.right_click.tags.title'),
+    icon: 'mdi-tag',
+    disabled: maindataStore.tags.length === 0,
+    disabledText: t('dashboard.right_click.tags.disabled_title'),
+    disabledIcon: 'mdi-tag-off',
+    children: maindataStore.tags.map(tag => ({
+      text: tag,
+      icon: hasTag(tag) ? 'mdi-checkbox-marked' : 'mdi-checkbox-blank-outline',
+      action: async () => await toggleTag(tag)
+    }))
+  },
+  {
+    text: t('dashboard.right_click.category.title'),
+    icon: 'mdi-label',
+    disabled: maindataStore.categories.length === 0,
+    disabledText: t('dashboard.right_click.category.disabled_title'),
+    disabledIcon: 'mdi-label-off',
+    children: availableCategories.value.map(category => ({
+      text: category.name === '' ? t('dashboard.right_click.category.clear') : category.name,
+      action: async () => await torrentStore.setTorrentCategory(hashes.value, category.name)
+    }))
+  },
+  {
+    text: t('dashboard.right_click.speed_limit.title'),
+    icon: 'mdi-speedometer-slow',
+    children: [
+      {
+        text: t('dashboard.right_click.speed_limit.download'),
+        icon: 'mdi-download',
+        action: setDownloadLimit
+      },
+      {
+        text: t('dashboard.right_click.speed_limit.upload'),
+        icon: 'mdi-upload',
+        action: setUploadLimit
+      },
+      {
+        text: t('dashboard.right_click.speed_limit.share'),
+        icon: 'mdi-account-group',
+        action: setShareLimit
+      }
+    ]
+  },
+  {
+    text: t('dashboard.right_click.copy.title'),
+    icon: 'mdi-content-copy',
+    hidden: isMultiple.value,
+    children: [
+      {
+        text: t('dashboard.right_click.copy.name'),
+        icon: 'mdi-alphabetical-variant',
+        action: async () => torrent.value && (await copyValue(torrent.value.name))
+      },
+      {
+        text: t('dashboard.right_click.copy.hash'),
+        icon: 'mdi-pound',
+        action: async () => await copyValue(hash.value)
+      },
+      {
+        text: t('dashboard.right_click.copy.magnet'),
+        icon: 'mdi-magnet',
+        action: async () => torrent.value && (await copyValue(torrent.value.magnet))
+      }
+    ]
+  },
+  {
+    text: t('dashboard.right_click.export', dashboardStore.selectedTorrents.length),
+    icon: isMultiple.value ? 'mdi-download-multiple' : 'mdi-download',
+    action: exportTorrents
+  },
+  {
+    text: t('dashboard.right_click.info'),
+    icon: 'mdi-information',
+    hidden: isMultiple.value,
+    action: () => router.push({ name: 'torrentDetail', params: { hash: hash.value } })
+  }
+])
+
 watch(
   () => trcProperties.isVisible,
   newValue => {
@@ -262,7 +530,41 @@ onBeforeUnmount(() => {
   </div>
 
   <div :style="`position: absolute; left: ${trcProperties.offset[0]}px; top: ${trcProperties.offset[1]}px;`">
-    <RightClickMenu v-model="trcProperties.isVisible" />
+    <RightClickMenu v-model="trcProperties.isVisible" :menu-data="menuData">
+      <template v-slot:top>
+        <v-list-item>
+          <div class="d-flex justify-space-around">
+            <v-tooltip location="top">
+              <template v-slot:activator="{ props }">
+                <v-btn density="compact" variant="plain" icon="mdi-play" v-bind="props" @click="resumeTorrents" />
+              </template>
+              <span>{{ $t('dashboard.right_click.top.resume') }}</span>
+            </v-tooltip>
+
+            <v-tooltip location="top">
+              <template v-slot:activator="{ props }">
+                <v-btn density="compact" variant="plain" icon="mdi-fast-forward" v-bind="props" @click="forceResumeTorrents" />
+              </template>
+              <span>{{ $t('dashboard.right_click.top.force_resume') }}</span>
+            </v-tooltip>
+
+            <v-tooltip location="top">
+              <template v-slot:activator="{ props }">
+                <v-btn density="compact" variant="plain" icon="mdi-pause" v-bind="props" @click="pauseTorrents" />
+              </template>
+              <span>{{ $t('dashboard.right_click.top.pause') }}</span>
+            </v-tooltip>
+
+            <v-tooltip location="top">
+              <template v-slot:activator="{ props }">
+                <v-btn color="red" density="compact" variant="plain" icon="mdi-delete-forever" v-bind="props" @click="deleteTorrents" />
+              </template>
+              <span>{{ $t('dashboard.right_click.top.delete') }}</span>
+            </v-tooltip>
+          </div>
+        </v-list-item>
+      </template>
+    </RightClickMenu>
   </div>
 </template>
